+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - #61 - Added new provider `Colt` using `ICal` and `Csv`
 - #66 - Added new provider `Momentum` using `Html` and `EmailSubjectParser`
 - #73 - Added new provider `Sparkle` using `Html` and `EmailSubjectParser`. Added support for multiple maintenances with `CombinedProcessor`.
+- #75 - Added new provider `AquaComms` using `Html` and `EmailSubjectParser`
 
 ### Fixed